enhancement/fonts: Update install font script
Adding message when font are already installed.

+++ scripts/install_font.sh
@@ -4,20 +4,29 @@ set -e
 
 # Configure font folder
 FONT_DIR="${HOME}/.local/share/fonts"
+UPDATE_CACHE=false
+
 mkdir -p ${FONT_DIR}
 cd ${FONT_DIR}
 
 function Download_Font {
     if [ $# -eq 1 ]
     then
-        mkdir ${1}
-        cd ${1}
+        if [ ! -d "${1}" ]
+        then
+            UPDATE_CACHE=true
+            mkdir ${1}
+            cd ${1}
 
-        wget -qO "${1}.zip" https://www.fontsquirrel.com/fonts/download/${1}
-        unzip ${1}.zip
-        rm -rf ${1}.zip
+            wget -qO "${1}.zip" https://www.fontsquirrel.com/fonts/download/${1}
+            unzip ${1}.zip
+            rm -rf ${1}.zip
 
-        cd ..
+            cd ..
+
+        else
+            echo "Using cached font ${1}"
+        fi
     fi
 }
 
@@ -34,4 +43,9 @@ Download_Font source-sans-pro
 Download_Font source-code-pro
 
 # Update cache
-fc-cache -frv
+if [ $UPDATE_CACHE = true ]
+then
+    fc-cache -frv
+else
+    echo "Cache no need to be updated"
+fi
